+++ src/app/setup/page.tsx
@@ -69,7 +69,7 @@ export default function SetupPage() {
   };
 
   return (
-    <div className="flex items-center justify-center min-h-screen bg-gray-100">
+    <div className="flex items-center justify-center min-h-screen">
       <div className="w-full max-w-md p-8 space-y-6 bg-white rounded-lg shadow-md">
         <h1 className="text-2xl font-bold text-center">Initial Setup</h1>
         <form onSubmit={handleSetup} className="space-y-6">
@@ -89,7 +89,7 @@ export default function SetupPage() {
                 setIsConnectionTested(false);
               }}
               required
-              className="w-full px-3 py-2 mt-1 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+              className="w-full px-3 py-2 mt-1"
             />
           </div>
           <div>
@@ -107,7 +107,7 @@ export default function SetupPage() {
                 setSecretKey(e.target.value);
                 setIsConnectionTested(false);
               }}
-              className="w-full px-3 py-2 mt-1 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+              className="w-full px-3 py-2 mt-1"
             />
           </div>
           <button
@@ -123,7 +123,7 @@ export default function SetupPage() {
               type="checkbox"
               checked={createAccount}
               onChange={(e) => setCreateAccount(e.target.checked)}
-              className="w-4 h-4 text-indigo-600 border-gray-300 rounded focus:ring-indigo-500"
+              className="w-4 h-4 text-primary border-gray-300 rounded focus:ring-primary"
             />
             <label
               htmlFor="createAccount"
@@ -147,7 +147,7 @@ export default function SetupPage() {
                   value={username}
                   onChange={(e) => setUsername(e.target.value)}
                   required={createAccount}
-                  className="w-full px-3 py-2 mt-1 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+                  className="w-full px-3 py-2 mt-1"
                 />
               </div>
               <div>
@@ -163,17 +163,21 @@ export default function SetupPage() {
                   value={password}
                   onChange={(e) => setPassword(e.target.value)}
                   required={createAccount}
-                  className="w-full px-3 py-2 mt-1 border border-gray-300 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500"
+                  className="w-full px-3 py-2 mt-1"
                 />
               </div>
             </>
           )}
           {error && <p className="text-sm text-red-600">{error}</p>}
-          {successMessage && <p className="text-sm text-green-600">{successMessage}</p>}
+          {successMessage && (
+            <p className="text-sm text-green-600">{successMessage}</p>
+          )}
           <button
             type="submit"
             disabled={!isConnectionTested}
-            className={`w-full px-4 py-2 font-medium text-white bg-indigo-600 rounded-md hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 ${!isConnectionTested ? 'opacity-50 cursor-not-allowed' : ''}`}
+            className={`w-full px-4 py-2 font-medium btn-primary ${
+              !isConnectionTested ? 'opacity-50 cursor-not-allowed' : ''
+            }`}
           >
             Save Configuration
           </button>
